+++ src/ModelManager/ThreadManagerInterface.php
@@ -18,17 +18,17 @@ interface ThreadManagerInterface extends ReadableManagerInterface
     /**
      * Finds a thread by its ID.
      */
-    public function findThreadById($id) : ?ThreadInterface;
+    public function findThreadById(int $id) : ?ThreadInterface;
 
     /**
      * Finds all threads in which a participant is involved.
      */
-    public function getParticipantThreadsQueryBuilder(ParticipantInterface $participant);
+    public function getParticipantThreadsQueryBuilder(int|ParticipantInterface $participant);
 
     /**
      * Finds in how many threads a participant is involved.
      */
-    public function getNbParticipantThreadsQueryBuilder(ParticipantInterface $participant);
+    public function getNbParticipantThreadsQueryBuilder(int|ParticipantInterface $participant);
 
     /**
      * Finds not deleted threads for a participant,
@@ -36,7 +36,7 @@ interface ThreadManagerInterface extends ReadableManagerInterface
      * ordered by last message not written by this participant in reverse order.
      * In one word: an inbox.
      */
-    public function getParticipantInboxThreadsQueryBuilder(ParticipantInterface $participant);
+    public function getParticipantInboxThreadsQueryBuilder(int|ParticipantInterface $participant);
 
     /**
      * Finds not deleted threads for a participant,
@@ -44,7 +44,7 @@ interface ThreadManagerInterface extends ReadableManagerInterface
      * ordered by last message not written by this participant in reverse order.
      * In one word: an inbox.
      */
-    public function findParticipantInboxThreads(ParticipantInterface $participant) : array;
+    public function findParticipantInboxThreads(int|ParticipantInterface $participant) : array;
 
     /**
      * Finds not deleted threads from a participant,
@@ -52,7 +52,7 @@ interface ThreadManagerInterface extends ReadableManagerInterface
      * ordered by last message written by this participant in reverse order.
      * In one word: an sentbox.
      */
-    public function getParticipantSentThreadsQueryBuilder(ParticipantInterface $participant);
+    public function getParticipantSentThreadsQueryBuilder(int|ParticipantInterface $participant);
 
     /**
      * Finds not deleted threads from a participant,
@@ -60,43 +60,43 @@ interface ThreadManagerInterface extends ReadableManagerInterface
      * ordered by last message written by this participant in reverse order.
      * In one word: an sentbox.
      */
-    public function findParticipantSentThreads(ParticipantInterface $participant) : array;
+    public function findParticipantSentThreads(int|ParticipantInterface $participant) : array;
 
     /**
      * Finds deleted threads from a participant,
      * ordered by last message date.
      */
-    public function getParticipantDeletedThreadsQueryBuilder(ParticipantInterface $participant);
+    public function getParticipantDeletedThreadsQueryBuilder(int|ParticipantInterface $participant);
 
     /**
      * Finds deleted threads from a participant,
      * ordered by last message date.
      */
-    public function findParticipantDeletedThreads(ParticipantInterface $participant) : array;
+    public function findParticipantDeletedThreads(int|ParticipantInterface $participant) : array;
 
     /**
      * Finds not deleted threads for a participant,
      * matching the given search term
      * ordered by last message not written by this participant in reverse order.
      */
-    public function getParticipantThreadsBySearchQueryBuilder(ParticipantInterface $participant, string $search);
+    public function getParticipantThreadsBySearchQueryBuilder(int|ParticipantInterface $participant, string $search);
 
     /**
      * Finds not deleted threads for a participant,
      * matching the given search term
      * ordered by last message not written by this participant in reverse order.
      */
-    public function findParticipantThreadsBySearch(ParticipantInterface $participant, string $search) : array;
+    public function findParticipantThreadsBySearch(int|ParticipantInterface $participant, string $search) : array;
 
     /**
      * Gets threads created by a participant.
      */
-    public function getThreadsCreatedByParticipantQueryBuilder(ParticipantInterface $participant);
+    public function getThreadsCreatedByParticipantQueryBuilder(int|ParticipantInterface $participant);
 
     /**
      * Gets threads created by a participant.
      */
-    public function findThreadsCreatedBy(ParticipantInterface $participant) : array;
+    public function findThreadsCreatedBy(int|ParticipantInterface $participant) : array;
 
     /**
      * Creates an empty comment thread instance.